Refreshing the data guard standby database
Notes on refreshing the data guard standby database. The standby database was out of sync with the primary database so I was compiling notes on how to refresh and get the standby back into sync with the primary database. These notes are not complete as I didn't have time on the contract I was on to complete the work. So in truth, I don't know how useful these notes are. If nothing else, it's a start.